magia a medio temrinar

+++ sample1/src/main/java/sample1/dao/EntityManagerMock.java
@@ -37,7 +37,7 @@ public class EntityManagerMock {
 	 * @param radio expresado en km ? FIXME
 	 * @return
 	 */
-	public List<Lugar> getLugaresXY(Integer x, Integer y , Integer radio){
+	public List<Lugar> getLugaresXY(Double x, Double y , Double radio){
 		return lugares.stream().filter(lugar -> HelperVarios.filtraDistancia(lugar,x,y,radio)).collect(Collectors.toList());
 	}
