+++ sample1/src/main/java/sample1/service/StockService.java
@@ -63,9 +63,9 @@ public class StockService {
 	 */
 	// que bares tienen growler para la ubicacion: growl + lugar a partir del:
 	// ubicacion.
-	public List<LugarStock> getLugaresGrowlerByUbicacion(Integer x, Integer y, List<Cerveza> tipoCervezasParsed) {
+	public List<LugarStock> getLugaresGrowlerByUbicacion(Double x, Double y, List<Cerveza> tipoCervezasParsed) {
 		
-		Integer radio = 5; // 1 km , TODO: ver de si lo paso por parametro o no.
+		Double radio = 5d; // 5 km , TODO: ver de si lo paso por parametro o no.
 		// dame los lugares cerca de x/y
 		List<Lugar> lugaresXY = emMock.getLugaresXY(x, y, radio);
 
@@ -85,9 +85,9 @@ public class StockService {
 	 */
 	// que bares tienen growler para la ubicacion: growl + lugar a partir del:
 	// ubicacion.
-	public List<LugarStock> getLugaresByUbicacion(Integer x, Integer y, List<Cerveza> tipoCervezasParsed) {
+	public List<LugarStock> getLugaresByUbicacion(Double x, Double y, List<Cerveza> tipoCervezasParsed) {
 		
-		Integer radio = 5; // 1 km , TODO: ver de si lo paso por parametro o no.
+		Double radio = 5d; // 1 km , TODO: ver de si lo paso por parametro o no.
 		// dame los lugares cerca de x/y
 		List<Lugar> lugaresXY = emMock.getLugaresXY(x, y, radio);
